Add files via upload
S.J. Sheather, A Modern Approach to Regression with R,
 3.4 Exercises
 1. The data file airfares.txt on the book web site gives the one-way airfare (in US dollars) and distance (in miles) from city A to 17 other cities in the US. Interest centers
on modeling airfare as a function of distance.

 3. The price of advertising (and hence revenue from advertising) is different from
one consumer magazine to another. Publishers of consumer magazines argue
that magazines that reach more readers create more value for the advertiser.
Thus, circulation is an important factor that affects revenue from advertising. In
this exercise, we are going to investigate the effect of circulation on gross
advertising revenue. The data are for the top 70 US magazines ranked in terms
of total gross advertising revenue in 2006. In particular we will develop regression models to predict gross advertising revenue per advertising page in 2006
(in thousands of dollars) from circulation (in millions). The data were obtained
from http://adage.com and are given in the file AdRevenue.csv which is available on the book web site. Prepare your answers to parts A, B and C in the form
of a report. 

 4. Tryfos (1998, p. 57) considers a real example involving the management at a
Canadian port on the Great Lakes who wish to estimate the relationship between
the volume of a ship’s cargo and the time required to load and unload this cargo. It
is envisaged that this relationship will be used for planning purposes as well as for
making comparisons with the productivity of other ports. Records of the tonnage
loaded and unloaded as well as the time spent in port by 31 liquid-carrying vessels
that used the port over the most recent summer are available. The data are available
on the book website in the file glakes.txt.